アカウント削除時に確認メッセージ #170
（送信時に無効化する処理は、Vue的に微妙だったのでいったんコメントアウト）
